refactor: useLoading으로 추상화
