+++ src/pages/Game/Game2/components/ResultSection/index.tsx
@@ -1,19 +1,11 @@
+import { useLoading } from '@/hooks/useLoading'
 import type { FC } from 'react'
-import { useEffect, useState } from 'react'
 import type { SectionProps } from '../../Game2.types'
 import { LoadingSection } from '../Loading'
 import * as S from './Result.styles'
 
 export const ResultSection: FC<SectionProps> = ({ onNextState }) => {
-  const [loading, setLoading] = useState(true)
-
-  useEffect(() => {
-    const loadingTimer = setTimeout(() => {
-      setLoading(false)
-    }, 3500)
-
-    return () => clearTimeout(loadingTimer)
-  }, [])
+  const { loading } = useLoading(3500)
 
   return <S.Container>{loading && <LoadingSection />}</S.Container>
 }
